revision

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -7,9 +7,9 @@ public class StatsServiceTest {
     @Test
     public void sumOfAllSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 180;
-        int actualSum = service.salesMonthSum(sales);
+        long actualSum = service.salesMonthSum(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
 
@@ -18,9 +18,9 @@ public class StatsServiceTest {
     @Test
     public void salesMonthAverageSum() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 15;
-        int actualSum = service.salesMonthAverageSum(sales);
+        long actualSum = service.salesMonthAverageSum(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
     }
@@ -28,9 +28,9 @@ public class StatsServiceTest {
     @Test
     public void monthWithMinimumSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 9;
-        int actualSum = service.minSales(sales);
+        long actualSum = service.minSales(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
     }
@@ -38,9 +38,9 @@ public class StatsServiceTest {
     @Test
     public void monthWithMaximumSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 8;
-        int actualSum = service.maxSales(sales);
+        long actualSum = service.maxSales(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
     }
@@ -48,9 +48,9 @@ public class StatsServiceTest {
     @Test
     public void numberMonthsBelowAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 5;
-        int actualSum = service.numberMonthsBelowAverage(sales);
+        long actualSum = service.numberMonthsBelowAverage(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
     }
@@ -58,9 +58,9 @@ public class StatsServiceTest {
     @Test
     public void numberMonthsAboveAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 5;
-        int actualSum = service.numberMonthsAboveAverage(sales);
+        long actualSum = service.numberMonthsAboveAverage(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
     }
